+++ scripts/storage.js
@@ -1,4 +1,4 @@
-var changes = 105; // Update this if readingPagesConfig is updated
+var changes = 106; // Update this if readingPagesConfig is updated
 
 var readingPagesConfig = {
 	readingConfigName: '',
@@ -171,6 +171,7 @@ var storageDefault = {
 		startOnlyFromLibrary: true,
 		startOnStartup: false,
 		ignoreSingleFoldersLibrary: true,
+		ignoreFilesRegex: '',
 		openingBehaviorFolder: 'file-list',
 		openingBehaviorFile: 'continue-reading-first-page',
 		useTheFirstImageAsPosterInFolders: false,